parquet2: parallelize meta reading.
