+++ src/query/storages/parquet/src/parquet2/parquet_table/mod.rs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod blocking_create;
-mod non_blocking_create;
+mod create;
 mod partition;
 mod read;
 mod table;
