+++ src/query/storages/parquet/src/parquet2/pruning.rs
@@ -19,7 +19,6 @@ use std::io::Seek;
 use std::sync::Arc;
 
 use common_arrow::arrow::datatypes::Field as ArrowField;
-use common_arrow::arrow::io::parquet::read as pread;
 use common_arrow::arrow::io::parquet::read::get_field_pages;
 use common_arrow::arrow::io::parquet::read::indexes::compute_page_row_intervals;
 use common_arrow::arrow::io::parquet::read::indexes::read_columns_indexes;
@@ -80,21 +79,6 @@ pub struct PartitionPruner {
     pub max_memory_usage: u64,
 }
 
-pub fn check_parquet_schema(
-    expect: &SchemaDescriptor,
-    actual: &SchemaDescriptor,
-    path: &str,
-    schema_from: &str,
-) -> Result<()> {
-    if expect.fields() != actual.fields() || expect.columns() != actual.columns() {
-        return Err(ErrorCode::BadBytes(format!(
-            "infer schema from '{}', but get diff schema in file '{}'. Expected schema: {:?}, actual: {:?}",
-            schema_from, path, expect, actual
-        )));
-    }
-    Ok(())
-}
-
 impl PartitionPruner {
     pub fn prune_one_file(
         &self,
@@ -118,12 +102,6 @@ impl PartitionPruner {
         file_meta: FileMetaData,
         operator: Operator,
     ) -> Result<(PartStatistics, Vec<Parquet2RowGroupPart>)> {
-        check_parquet_schema(
-            &self.schema_descr,
-            file_meta.schema(),
-            path,
-            &self.schema_from,
-        )?;
         let mut stats = PartStatistics::default();
         let mut partitions = vec![];
 
@@ -232,6 +210,7 @@ impl PartitionPruner {
         &self,
         operator: Operator,
         locations: &Vec<(String, u64)>,
+        num_threads: usize,
     ) -> Result<(PartStatistics, Partitions)> {
         // part stats
         let mut stats = PartStatistics::default();
@@ -248,32 +227,14 @@ impl PartitionPruner {
 
         let mut partitions = Vec::with_capacity(locations.len());
 
-        let is_blocking_io = operator.info().can_blocking();
-
         // 1. Read parquet meta data. Distinguish between sync and async reading.
-        let file_metas = if is_blocking_io {
-            let mut file_metas = Vec::with_capacity(locations.len());
-            for (location, _size) in &large_files {
-                let mut reader = operator.blocking().reader(location)?;
-                let file_meta = pread::read_metadata(&mut reader).map_err(|e| {
-                    ErrorCode::Internal(format!(
-                        "Read parquet file '{}''s meta error: {}",
-                        location, e
-                    ))
-                })?;
-                file_metas.push(file_meta);
-            }
-            file_metas
-        } else {
-            read_parquet_metas_in_parallel(
-                operator.clone(),
-                large_files.clone(),
-                16,
-                64,
-                self.max_memory_usage,
-            )
-            .await?
-        };
+        let file_metas = read_parquet_metas_in_parallel(
+            operator.clone(),
+            large_files.clone(),
+            num_threads,
+            self.max_memory_usage,
+        )
+        .await?;
 
         // 2. Use file meta to prune row groups or pages.
         let mut max_compression_ratio = self.compression_ratio;
